backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
supabase>=2.3.0
pytesseract>=0.3.10
Pillow>=10.3.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
opencv-python>=4.9.0.80
numpy>=1.26.0
